requirements.txt
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
openai==1.58.1
djangorestframework==3.15.2